pyproject.toml
[project]
name = "agent-gm"
version = "0.1.0"
description = "Multi-agent text game dungeon master"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "strands-agents[litellm]>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "sqlalchemy>=2.0.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "streamlit>=1.30.0",
    "pandas>=2.0.0",
    "plotly>=6.5.0",
]

[project.scripts]
agent-gm = "main:main"
